just the ViewController.swift from Project25
The one thing I did differently from the instructions from @twostraws was there was an error shown when I added the Protocols - when I said "Fixit" it added stubs for 7 methods which I thought was really cool. Then I watched the rest of the video as he filled them all in (or not) Maybe I messed something up there.